moved implementations into headers together with their declaration
