+++ .gitignore
@@ -15,3 +15,5 @@ CMakeUserPresets.json
 *.pyc
 __pycache__
 *.pyd
+.venv
+venv
