+++ cudax/include/cuda/experimental/__cufile/buffer_handle.hpp
@@ -102,6 +102,93 @@ public:
   bool is_valid() const noexcept;
 };
 
-} // namespace cuda::experimental::cufile
+// ===================== Inline implementations =====================
 
-#include <cuda/experimental/__cufile/detail/buffer_handle_impl.hpp>
+// Constructor implementations
+template <typename T>
+inline buffer_handle::buffer_handle(::cuda::std::span<T> buffer, int flags)
+    : buffer_(::cuda::std::as_bytes(buffer))
+{
+  static_assert(::cuda::std::is_trivially_copyable_v<T>, "Type must be trivially copyable for cuFile operations");
+
+  CUfileError_t error = cuFileBufRegister(buffer_.data(), buffer_.size(), flags);
+  detail::check_cufile_result(error, "cuFileBufRegister");
+
+  registered_buffer_.emplace(buffer_.data(), [](const void* buf) {
+    cuFileBufDeregister(buf);
+  });
+}
+
+template <typename T>
+inline buffer_handle::buffer_handle(::cuda::std::span<const T> buffer, int flags)
+    : buffer_(::cuda::std::as_bytes(buffer))
+{
+  static_assert(::cuda::std::is_trivially_copyable_v<T>, "Type must be trivially copyable for cuFile operations");
+
+  CUfileError_t error = cuFileBufRegister(buffer_.data(), buffer_.size(), flags);
+  detail::check_cufile_result(error, "cuFileBufRegister");
+
+  registered_buffer_.emplace(buffer_.data(), [](const void* buf) {
+    cuFileBufDeregister(buf);
+  });
+}
+
+// Move constructor and assignment
+inline buffer_handle::buffer_handle(buffer_handle&& other) noexcept
+    : buffer_(other.buffer_)
+    , registered_buffer_(::std::move(other.registered_buffer_))
+{}
+
+inline buffer_handle& buffer_handle::operator=(buffer_handle&& other) noexcept
+{
+  if (this != &other)
+  {
+    buffer_            = other.buffer_;
+    registered_buffer_ = ::std::move(other.registered_buffer_);
+  }
+  return *this;
+}
+
+// Simple getter implementations
+inline const void* buffer_handle::data() const noexcept
+{
+  return buffer_.data();
+}
+
+inline size_t buffer_handle::size() const noexcept
+{
+  return buffer_.size();
+}
+
+inline ::cuda::std::span<const ::cuda::std::byte> buffer_handle::as_bytes() const noexcept
+{
+  return buffer_;
+}
+
+inline ::cuda::std::span<::cuda::std::byte> buffer_handle::as_writable_bytes() const noexcept
+{
+  return ::cuda::std::span<::cuda::std::byte>(const_cast<::cuda::std::byte*>(buffer_.data()), buffer_.size());
+}
+
+// Template method implementations
+template <typename T>
+inline ::cuda::std::span<T> buffer_handle::as_span() const noexcept
+{
+  static_assert(::cuda::std::is_trivially_copyable_v<T>, "Type must be trivially copyable for cuFile operations");
+  return ::cuda::std::span<T>(
+    reinterpret_cast<T*>(const_cast<::cuda::std::byte*>(buffer_.data())), buffer_.size() / sizeof(T));
+}
+
+template <typename T>
+inline ::cuda::std::span<const T> buffer_handle::as_const_span() const noexcept
+{
+  static_assert(::cuda::std::is_trivially_copyable_v<T>, "Type must be trivially copyable for cuFile operations");
+  return ::cuda::std::span<const T>(reinterpret_cast<const T*>(buffer_.data()), buffer_.size() / sizeof(T));
+}
+
+inline bool buffer_handle::is_valid() const noexcept
+{
+  return registered_buffer_.has_value();
+}
+
+} // namespace cuda::experimental::cufile
